Replaced MoveOptionalUndo with OptionalUndoEdit wrapper.
The wrapper allows for adding optional undo to any
edit. This is a workaround for allowing small
incremental edits that is not pushed to the
undo stack and then push a edit that is not executed
but are pushed to the undo stack representing the
operations done by all smaller incremental edits.