A bit matrix which will help in minimisation when constructing cycle bases
Utility class for manipulating int[][] adjacency list graphs

Allow null containers in the ShortestPaths tool

Removing toAdjList from ShortestPaths and RingSearch the new utility class is now used

Allowing internal uses to specify an ordering for the ShortestPaths tool

A tool to discover the set of initial cycles. This super-set can then be used to extract the minimum cycle basis and relevant cycles

Extraction of a cycle basis using a greedy algorithm

Concise algorithm for relevant cycles

Including new test classes in the module suite

Minimum Cycle Basis implementation reusing the set of initial cycles

Small tweaks to the BitMatrix, elimination now returns the rank of the matrix directly.

Utility methods for creating the BitMatrix from cycle collections.

Using the new matrix utility and added a method to access the basis size directly.

The essential cycles of a graph.

Updating documentation of MCB